+++ packages/amplify-util-mock/src/__tests__/utils/lambda/populate-cfn-params.test.ts
@@ -1,6 +1,7 @@
 import { stateManager } from '@aws-amplify/amplify-cli-core';
 import _ from 'lodash';
 import { populateCfnParams } from '../../../utils/lambda/populate-cfn-params';
+import { printer } from '@aws-amplify/amplify-prompts';
 
 jest.mock('@aws-amplify/amplify-cli-core');
 jest.mock('../../../api/api', () => ({
@@ -19,6 +20,15 @@ stateManager_mock.getLocalEnvInfo.mockReturnValue({
 const teamProviderParam = {
   anotherCfnParam: 'testValue',
 };
+stateManager_mock.getMeta.mockReturnValue({
+  providers: {
+    awscloudformation: {
+      Region: 'test-region',
+      StackId: 'arn:aws:cloudformation:us-test-1:1234:stack/my-test-stack',
+      StackName: 'test-stack-name',
+    },
+  },
+});
 stateManager_mock.getTeamProviderInfo.mockReturnValue({
   test: {
     awscloudformation: {
@@ -69,9 +79,18 @@ const meta_stub = {
   },
 };
 
+let ensureEnvParamManager;
 describe('populate cfn params', () => {
-  it('includes CFN pseudo parameters', () => {
-    expect(populateCfnParams({} as any, undefined)).toMatchObject({
+  beforeAll(async () => {
+    ({ ensureEnvParamManager } = await import('@aws-amplify/amplify-environment-parameters'));
+    await ensureEnvParamManager('test');
+  });
+
+  afterEach(async () => {
+    jest.clearAllMocks();
+  });
+  it('includes CFN pseudo parameters', async () => {
+    expect(populateCfnParams('test')).toMatchObject({
       env: 'test',
       'AWS::Region': 'test-region',
       'AWS::AccountId': '1234',
@@ -81,14 +100,13 @@ describe('populate cfn params', () => {
     });
   });
 
-  it('falls back to default pseudo params when not in team provider', () => {
-    stateManager_mock.getTeamProviderInfo.mockReturnValueOnce({
-      test: {
+  it('falls back to default pseudo params when not in team provider', async () => {
+    stateManager_mock.getMeta.mockReturnValueOnce({
+      providers: {
         awscloudformation: {},
       },
     });
-
-    expect(populateCfnParams({} as any, undefined)).toMatchObject({
+    expect(populateCfnParams('test')).toMatchObject({
       env: 'test',
       'AWS::Region': 'us-test-1',
       'AWS::AccountId': '12345678910',
@@ -98,62 +116,62 @@ describe('populate cfn params', () => {
     });
   });
 
-  it('gets dependsOn params from amplify-meta', () => {
-    stateManager_mock.getMeta.mockReturnValueOnce(meta_stub);
-    expect(populateCfnParams({} as any, 'func1')).toMatchObject({
+  it('gets dependsOn params from amplify-meta', async () => {
+    stateManager_mock.getMeta.mockReturnValue(meta_stub);
+    expect(populateCfnParams('func1')).toMatchObject({
       apimyApiapiName: 'testApiName',
       storagemytabletableName: 'testTableName',
       storagemytabletableArn: 'testTableArn',
     });
   });
 
-  it('overwrites api endpoint url when specified', () => {
+  it('overwrites api endpoint url when specified', async () => {
     const meta_stub_copy = _.cloneDeep(meta_stub);
     meta_stub_copy.function.func1.dependsOn[1].attributes.push('GraphQLAPIEndpointOutput');
-    stateManager_mock.getMeta.mockReturnValueOnce(meta_stub_copy);
-    expect(populateCfnParams({ warning: jest.fn() } as any, 'func1', true)).toMatchObject({
+    stateManager_mock.getMeta.mockReturnValue(meta_stub_copy);
+    expect(populateCfnParams('func1', true)).toMatchObject({
       apimyApiGraphQLAPIEndpointOutput: `http://localhost:666/graphql`,
     });
   });
 
-  it('overwrites api key when specified', () => {
+  it('overwrites api key when specified', async () => {
     const meta_stub_copy = _.cloneDeep(meta_stub);
     meta_stub_copy.function.func1.dependsOn[1].attributes.push('GraphQLAPIKeyOutput');
-    stateManager_mock.getMeta.mockReturnValueOnce(meta_stub_copy);
-    expect(populateCfnParams({ warning: jest.fn() } as any, 'func1', true)).toMatchObject({
+    stateManager_mock.getMeta.mockReturnValue(meta_stub_copy);
+    expect(populateCfnParams('func1', true)).toMatchObject({
       apimyApiGraphQLAPIKeyOutput: 'da2-fakeApiId123456',
     });
   });
 
-  it('prints warning when no value found', () => {
+  it('prints warning when no value found', async () => {
     const meta_stub_copy = _.cloneDeep(meta_stub);
+    stateManager_mock.getMeta.mockReturnValue(meta_stub_copy);
     meta_stub_copy.function.func1.dependsOn[1].attributes.push('GraphQLAPIEndpointOutput');
-    stateManager_mock.getMeta.mockReturnValueOnce(meta_stub_copy);
-    const warningMock = jest.fn();
-    const result = populateCfnParams({ warning: warningMock } as any, 'func1');
+    const warningSpy = jest.spyOn(printer, 'warn');
+    const result = populateCfnParams('func1');
     expect(typeof result).toBe('object');
     expect(result.apimyApiGraphQLAPIEndpointOutput).toBeUndefined();
-    expect(warningMock.mock.calls).toMatchInlineSnapshot(`
-      Array [
-        Array [
-          "No output found for attribute 'GraphQLAPIEndpointOutput' on resource 'myApi' in category 'api'",
-        ],
-        Array [
-          "This attribute will be undefined in the mock environment until you run \`amplify push\`",
-        ],
-      ]
-    `);
+    expect(warningSpy.mock.calls).toMatchInlineSnapshot(`
+    [
+      [
+        "No output found for attribute 'GraphQLAPIEndpointOutput' on resource 'myApi' in category 'api'",
+      ],
+      [
+        "This attribute will be undefined in the mock environment until you run \`amplify push\`",
+      ],
+    ]
+  `);
   });
 
-  it('includes params from parameters.json', () => {
+  it('includes params from parameters.json', async () => {
     const expectedMap = {
       someOtherParam: 'this is the value',
     };
     stateManager_mock.getResourceParametersJson.mockReturnValueOnce(expectedMap);
-    expect(populateCfnParams({} as any, 'func1')).toMatchObject(expectedMap);
+    expect(populateCfnParams('func1')).toMatchObject(expectedMap);
   });
 
   it('includes params from team-provider-info.json', () => {
-    expect(populateCfnParams({} as any, 'func1')).toMatchObject(teamProviderParam);
+    expect(populateCfnParams('func1')).toMatchObject(teamProviderParam);
   });
 });